Fix subscription mechanism

+++ backend/api/views.py
@@ -51,10 +51,19 @@ class CustomUserViewSet(UserViewSet):
         user = request.user
         author_id = kwargs.get('id')
         author = get_object_or_404(User, id=author_id)
+        subscription = Subscription.objects.filter(
+            user=user,
+            author=author
+        )
         if request.method == 'POST':
             if user == author:
                 return Response(
                     {'errors': 'It\'s not allowed to subscribe to yourself.'},
+                    status=status.HTTP_400_BAD_REQUEST
+                )
+            elif subscription.exists():
+                return Response(
+                    {'errors': 'You are already subscribed to the user'},
                     status=status.HTTP_400_BAD_REQUEST
                 )
             serializer = SubscriptionSerializer(
@@ -67,27 +76,33 @@ class CustomUserViewSet(UserViewSet):
             return Response(
                 serializer.data, status=status.HTTP_201_CREATED
             )
-        subscription = get_object_or_404(
-            Subscription,
-            user=user,
-            author=author
-        )
+        if not subscription.exists():
+            return Response(
+                {'errors': 'You are not subscribed to the user'},
+                status=status.HTTP_400_BAD_REQUEST
+            )
         subscription.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
     @action(
         detail=False,
         permission_classes=[IsAuthenticated],
+        methods=['get']
     )
     def subscriptions(self, request):
         queryset = User.objects.filter(
             subscriber__user=request.user
         )
         page = self.paginate_queryset(queryset)
+        if page is not None:
+            serializer = SubscriptionSerializer(
+                page, many=True, context={'request': request}
+            )
+            return self.get_paginated_response(serializer.data)
         serializer = SubscriptionSerializer(
-            page, many=True, context={'request': request}
+            queryset, many=True
         )
-        return self.get_paginated_response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
 
 class RecipeViewSet(ModelViewSet):
